+++ oap-server/server-alarm-plugin/src/main/java/org/apache/skywalking/oap/server/core/alarm/provider/NotifyHandler.java
@@ -41,15 +41,18 @@ import org.apache.skywalking.oap.server.core.analysis.metrics.Metrics;
 import org.apache.skywalking.oap.server.core.analysis.metrics.MetricsMetaInfo;
 import org.apache.skywalking.oap.server.core.analysis.metrics.WithMetadata;
 import org.apache.skywalking.oap.server.core.source.DefaultScopeDefine;
+import org.apache.skywalking.oap.server.library.module.ModuleManager;
 
 @Slf4j
 public class NotifyHandler implements MetricsNotify {
     private final AlarmCore core;
     private final AlarmRulesWatcher alarmRulesWatcher;
+    private final ModuleManager manager;
 
-    public NotifyHandler(AlarmRulesWatcher alarmRulesWatcher) {
+    public NotifyHandler(AlarmRulesWatcher alarmRulesWatcher, ModuleManager manager) {
         this.alarmRulesWatcher = alarmRulesWatcher;
         core = new AlarmCore(alarmRulesWatcher);
+        this.manager = manager;
     }
 
     @Override
@@ -166,6 +169,7 @@ public class NotifyHandler implements MetricsNotify {
         allCallbacks.add(new WechatHookCallback(alarmRulesWatcher));
         allCallbacks.add(new DingtalkHookCallback(alarmRulesWatcher));
         allCallbacks.add(new FeishuHookCallback(alarmRulesWatcher));
+        allCallbacks.add(new EventHookCallback(this.manager));
         core.start(allCallbacks);
     }
 }