+++ CHANGES.md
@@ -19,6 +19,7 @@ Release Notes.
 #### OAP-Backend
 * BugFix: filter invalid Envoy access logs whose socket address is empty.
 * Fix K8s monitoring the incorrect metrics calculate. 
+* Loop alarm into event system.
 
 #### UI
 * Add logo for kong plugin.